Split PnL chart component
